requirements.txt
os
dotenv
numpy
pandas
langchain
streamlit
google-generativeai